+++ geoprox-server/src/handlers.rs
@@ -347,8 +347,8 @@ pub mod geoshard_api {
             &index,
             [query.lat, query.lng],
             query.range.into(),
-            query.count.unwrap_or(100),
-            query.sorted.unwrap_or(false),
+            query.count,
+            query.sorted,
         ) {
             Ok(found) => Ok(Json(QueryRangeResponse { found })),
             Err(err) => Err(anyhow!(err).into()),
@@ -380,8 +380,8 @@ pub mod geoshard_api {
             query.indices.into_iter().collect(),
             [query.lat, query.lng],
             query.range.into(),
-            query.count.unwrap_or(100),
-            query.sorted.unwrap_or(false),
+            query.count,
+            query.sorted,
         );
 
         Ok(Json(QueryRangeManyResponse {